% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{forestplot}
\alias{forestplot}
\title{Produce forest plot from netmeta object.}
\usage{
forestplot(data, file)
}
\arguments{
\item{data}{A netmeta object}

\item{file}{Filepath where the plot will be saved}
}
\description{
Code taken directly from Akter et al. Used in Figures 2 and 3.
Output is saved directly to a PDF file.
}
