% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data}
\alias{load_data}
\title{Load the study's data.}
\usage{
load_data(measure = NULL)
}
\arguments{
\item{measure}{Must exactly match "OR" or "MD".}
}
\value{
A tibble with data for the specified measure.
}
\description{
Load the study's data.
}
\references{
Akter, S., Rahman, M.M., Rouyard, T. et al. A systematic review and network
meta-analysis of population-level interventions to tackle smoking behaviour.
Nat Hum Behav 8, 2367–2391 (2024). \url{https://doi.org/10.1038/s41562-024-02002-7}
}
