% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{make_fig2}
\alias{make_fig2}
\title{Generate subfigures for Figure 2.}
\usage{
make_fig2()
}
\value{
Invisibly returns the path where the figures were saved.
}
\description{
Running this function will generate eight PDF files in the current working
directory (check with \code{getwd()}).
}
