% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figures.R
\name{make_fig3}
\alias{make_fig3}
\title{Generate subfigures for Figure 3.}
\usage{
make_fig3()
}
\value{
Invisibly returns the path where the figures were saved.
}
\description{
Running this function will generate four PDF files in the current working
directory (check with \code{getwd()}).
}
