% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{metaMD}
\alias{metaMD}
\title{Wrapper for the meta-analysis function called on MD measures.}
\usage{
metaMD(data)
}
\arguments{
\item{data}{A tibble with MD data (see \code{load_data}).}
}
\value{
A \code{meta} object.
}
\description{
Argument names changed according to newer version of \code{meta} package.
Results are unchanged.
}
