% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{metaOR}
\alias{metaOR}
\title{Wrapper for the meta-analysis function called on OR measures.}
\usage{
metaOR(data)
}
\arguments{
\item{data}{A tibble with OR data (see \code{load_data}).}
}
\value{
A \code{meta} object.
}
\description{
Argument names changed according to newer version of \code{meta} package.
Results are unchanged.
}
