% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta.R
\name{netmetaOR}
\alias{netmetaOR}
\title{Wrapper for the network meta-analysis function.}
\usage{
netmetaOR(data)
}
\arguments{
\item{data}{A tibble with OR data (see \code{load_data}).}
}
\value{
A \code{netmeta} object.
}
\description{
Used for network meta-analysis in Figures 2 and 3.
}
\details{
Argument names changed according to newer version of \code{netmeta} package.
Results are unchanged.
}
