% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{rankplot}
\alias{rankplot}
\title{Produce rank plot from netrank object.}
\usage{
rankplot(data)
}
\arguments{
\item{data}{A netrank object}
}
\description{
Code taken directly from Akter et al. Used in Figures 2 and 3.
}
