% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_data}
\alias{load_data}
\title{Load the study's data.}
\usage{
load_data(measure = NULL)
}
\arguments{
\item{measure}{Must exactly match "OR" or "MD".}
}
\value{
A tibble with data for the specified measure.
}
\description{
Load the study's data.
}
