% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{make_table4}
\alias{make_table4}
\title{Produce output for Table 4.}
\usage{
make_table4()
}
\value{
A tibble with the results of pairwise meta-analysis for
smoking consumption outcomes
}
\description{
All measures in the OR dataset are used in the calculation. This includes
multiple measures from the same study, as this was found to give the closest
match to the original results.
}
\details{
Cigarette consumption contains additional row not present in original table.
Number of studies is separately calculated based on the unique study IDs.
Row order adjusted manually to match original table.
}
